Refactoring: move POM_XML from Maven to Constants
