+++ src/test/java/tech/jhipster/lite/generator/server/sonar/application/SonarApplicationServiceIT.java
@@ -1,7 +1,7 @@
 package tech.jhipster.lite.generator.server.sonar.application;
 
 import static tech.jhipster.lite.TestUtils.*;
-import static tech.jhipster.lite.generator.buildtool.maven.domain.MavenDomainService.POM_XML;
+import static tech.jhipster.lite.generator.project.domain.Constants.POM_XML;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
